Use one parsingHook for SQL and DistSQL parser to sent tracing event (#7968)
* Use one parsingHook for SQL and DistSQL parser to sent tracing event

* Move ParsingHook to infra-parser module
